+++ src/apis/core/index.ts
@@ -36,7 +36,7 @@ interface CustomInstance extends AxiosInstance {
 }
 
 export const request: CustomInstance = axios.create({
-	baseURL: process.env.REACT_APP_API,
+	baseURL: import.meta.env.VITE_API_URL,
 	timeout: 20000,
 	headers: {
 		accept: 'application/json',
