+++ src/components/Cards/Card/dto.ts
