+++ src/components/Cards/Card/styles.tsx
